Fix Int32 overflow on prompt_eval_duration + add release workflow (v2.0.1) (#9)
* Fix Int32 overflow on prompt_eval_duration (v2.0.1)

Ollama reports durations in nanoseconds. Int32 tops out at ~2.147 seconds, so
ChatResponse.PromptEvalDuration and GenerateResponse.PromptEvalDuration — typed
int? while every other duration on those types was long? — threw

  JsonException: The JSON value could not be converted to
  System.Nullable`1[System.Int32]. Path: $.prompt_eval_duration

on any prompt that took longer than that to evaluate. A warm model stayed under
the limit, so this looked intermittent; a cold model overflowed it every time and
the entire response was discarded even though the model had answered correctly.

Both are now long?. Adds DurationOverflowTests covering the int.MaxValue boundary.


* Add release workflow: tag v* -> pack -> NuGet -> GitHub release

Mirrors the pipeline InferHub.Client already uses. The repo could pack but had no
way to publish, so releases were manual.

The tag is authoritative for the version: PackageVersion is set in the csproj, so
/p:Version alone would not win and it has to be overridden explicitly too.

Requires NUGET_API_KEY (repo secret, or the "Action" environment as in InferHub.Client).


---------

+++ src/OllamaClient/Models/ChatResponse.cs
@@ -61,11 +61,13 @@ public class ChatResponse
     public int? PromptEvalCount { get; set; }
 
     /// <summary>
-    /// prompt_eval_duration: time spent in nanoseconds evaluating the prompt
+    /// prompt_eval_duration: time spent in nanoseconds evaluating the prompt.
+    /// Nanoseconds overflow Int32 after ~2.1 seconds, so this must stay a long —
+    /// as int? it threw a JsonException on any prompt that took longer than that.
     /// </summary>
     [JsonPropertyName("prompt_eval_duration")]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public int? PromptEvalDuration { get; set; }
+    public long? PromptEvalDuration { get; set; }
 
     /// <summary>
     /// eval_count: number of tokens the response
